pyproject.toml
[project]
name = "ssh-relay"
version = "0.1.0"
description = "nassh-compatible SSH relay, gated by Cloudflare Access"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "pyjwt[crypto]>=2.9",
    "pydantic-settings>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-timeout>=2.3",
    "websockets>=13",
    "ruff>=0.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
timeout = 30

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ssh_relay"]

[tool.ruff]
line-length = 100
target-version = "py311"
